Frontend has been migrated Svelte, also, some backend
Login front and back finished. If anything I need to add something to prevent non-logged users to move to the home page, also, any error that may occur during authorization and requires user to stay at login page needs to be addressed. As for the home, it already pulls the user's personal info and playlists.